requirements/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
frozenlist==1.4.1
idna==3.6
marshmallow==3.21.1
multidict==6.0.5
packaging==24.0
PyYAML==6.0.1
webargs==8.4.0
yarl==1.9.4
